adicionei uma opç pra escolher se quer excluir

+++ 01-face_points_exp/error_verifier.py
@@ -2,6 +2,12 @@
 import os
 
 import shutil
+
+delete = input("Desseja deletar as inconsistencias detectadas? s/n   ")
+delete = delete == "s" or delete == "S"
+num_incorreto = 0
+pasta_vazia = 0
+com_erro = 0
 expressions = os.listdir("../processed")
 for exp in expressions:
     types = os.listdir("../processed/"+exp)
@@ -11,13 +17,17 @@ for exp in expressions:
             infos = os.listdir("../processed/"+exp+"/"+tp+"/"+user)
             if(len(infos) == 0):
                 print("Removendo pasta vazia: "+exp+"/"+tp+"/"+user)
-                shutil.rmtree("../processed/"+exp+"/"+tp+"/"+user)
+                if(delete):
+                    shutil.rmtree("../processed/"+exp+"/"+tp+"/"+user)
+                pasta_vazia += 1
                 continue
 
             if(len(infos) != 3):
                 print("Removendo pasta com numero incorreto de imagens: " +
                       exp+"/"+tp+"/"+user)
-                shutil.rmtree("../processed/"+exp+"/"+tp+"/"+user)
+                if(delete):
+                    shutil.rmtree("../processed/"+exp+"/"+tp+"/"+user)
+                num_incorreto += 1
                 continue
             achou = False
             for nm in infos:
@@ -26,5 +36,14 @@ for exp in expressions:
                     break
             if(achou):
                 print("Removendo pasta com erros: "+exp+"/"+tp+"/"+user)
-                shutil.rmtree("../processed/"+exp+"/"+tp+"/"+user)
+                com_erro += 1
+                if(delete):
+                    shutil.rmtree("../processed/"+exp+"/"+tp+"/"+user)
                 continue
+
+print()
+print()
+print()
+print("Pastas com numero de fotos incorreto: ", num_incorreto)
+print("Pastas Vazias: ", pasta_vazia)
+print("Pastas com erro: ", com_erro)
